Offer xPack-provided toolchain and OpenOCD packages for mac-arm64
This should improve the development experience on mac-arm64 systems
(Apple Silicon, M1 and later CPUs) by using toolchain and OpenOCD builds
for that architecture, rather than relying on binary translation of the
mac-x86_64 tools.

mac-arm64 tools are only added to the current versions of these packages
in this index: xpack-arm-none-eabi-gcc 13.2.1-1.1, and xpack-openocd
0.12.0-4, both as used in the current STM32 platform version, 2.9.0.
Future versions should continue to offer mac-arm64 builds in this
package index.

Link: #72
Signed-off-by: Mark Mentovai <mark@mentovai.com>